+++ trunk/src/edr/src/plugins/newVdf/src/StyleSinkNode.h
@@ -33,7 +33,7 @@ public:
     virtual QGraphicsItem * visualItem() const;
     virtual void setSelection(bool selected);
 
-	virtual int getNumInputPins() const { return pins.size(); }
+	virtual int getNumInputPins() const { return static_cast<int>(pins.size()); }
 	virtual IVisualInputPinPtr getInputPin(int no) { return pins[no]; }
 
 	virtual void setVisualStrategy( IVisualStrategyPtr strategy );
